refactor(xray-lite): move Namespace to namespace
Moves `Namespace` and implementers to a new submodule `namespace`.
